Removed code from index moved to component/LandingPage
